release: 2.0.0
feat: Added the Effect class which provides the particle effects.
feat: Integrated the BallTable which stores the Ball positions when they collide.
feat: Balls and the movable cannon are separated in new classes.
feat: SObject class is born to handle all place-, draw- and update able objects.
feat: First drop able Balls on a background Image and a Timer which counts the time after the game was started.