release: 3.1.0
update: Particle are now removed automatically -no more heap size errors.
fix: Added the five balls check (more than four balls on a stack, shrink to one heavy ball).
update: Complete rewrite of the 3 Balls detector and added an ability to remove also the connected balls.
feat: Added the HighScore Counter and the score counter.
feat: Added a program icon to the game and the executable.
feat: Added the Weight Table which displays the weight sum of the balls on each column.
feat: Added the LevelBall which displays the current level. Now every 50 balls a new color will be added.